refactor(mcp): relocate tool result envelope into core
core/tool_middleware.short_circuit_error imported
tools/tool_error_envelope, a back-edge against the inward-only layering
where tools sits above core. The error/success envelope is the shared
result currency used both by the tool adapters and by the core
middleware, so it belongs at core, not in tools.

Move the module to core/tool_error_envelope.py (and its test to
tests/core/), repoint every importer, and hoist the previously lazy
import in short_circuit_error now that relocating the module removes the
load-time cycle that forced it. No behavior change: the module content
is untouched.

This clears the pipefy_mcp.core -> pipefy_mcp.tools edge so the
intra-package layering contract holds.
