+++ packages/mcp/src/pipefy_mcp/core/tool_error_envelope.py
