+++ packages/mcp/src/pipefy_mcp/tools/observability_tool_helpers.py
@@ -6,10 +6,10 @@ from typing import Any, Literal, cast
 
 from typing_extensions import TypedDict
 
+from pipefy_mcp.core.tool_error_envelope import tool_error
 from pipefy_mcp.tools.graphql_error_helpers import (
     handle_tool_graphql_error,
 )
-from pipefy_mcp.tools.tool_error_envelope import tool_error
 
 
 class ObservabilityReadSuccessPayload(TypedDict):
