+++ packages/mcp/src/pipefy_mcp/tools/portal_tools.py
@@ -16,6 +16,7 @@ from pipefy_sdk.models.portal import (
 )
 from pydantic import ValidationError
 
+from pipefy_mcp.core.tool_error_envelope import tool_error
 from pipefy_mcp.tools.destructive_tool_guard import check_destructive_confirmation
 from pipefy_mcp.tools.introspection_tool_helpers import (
     build_error_payload,
@@ -32,7 +33,6 @@ from pipefy_mcp.tools.portal_tool_helpers import (
     validate_tool_ids,
 )
 from pipefy_mcp.tools.tool_context import get_pipefy_client
-from pipefy_mcp.tools.tool_error_envelope import tool_error
 from pipefy_mcp.tools.validation_helpers import validate_tool_id
