+++ packages/mcp/tests/tools/test_ai_agent_tools.py
@@ -19,8 +19,8 @@ from mcp.shared.memory import (
 )
 from pipefy_sdk.models.ai_agent import UpdateAiAgentInput
 
+from pipefy_mcp.core.tool_error_envelope import tool_error_message
 from pipefy_mcp.tools.ai_agent_tools import AiAgentTools
-from pipefy_mcp.tools.tool_error_envelope import tool_error_message
 from tools.conftest import (
     assert_invalid_arguments_envelope,
     build_tool_test_server,
