+++ packages/mcp/tests/tools/test_attachment_tools.py
@@ -18,8 +18,8 @@ from pipefy_sdk import (
 )
 from pipefy_sdk.models.attachment import infer_content_type
 
+from pipefy_mcp.core.tool_error_envelope import tool_error_message
 from pipefy_mcp.tools.attachment_tools import AttachmentTools
-from pipefy_mcp.tools.tool_error_envelope import tool_error_message
 from tools.conftest import build_tool_test_server
