+++ packages/mcp/tests/tools/test_field_condition_tools.py
@@ -10,8 +10,8 @@ from mcp.shared.memory import (
 )
 from pipefy_sdk import PipefyClient
 
+from pipefy_mcp.core.tool_error_envelope import tool_error_message
 from pipefy_mcp.tools.field_condition_tools import FieldConditionTools
-from pipefy_mcp.tools.tool_error_envelope import tool_error_message
 from tools.conftest import build_tool_test_server
