+++ packages/mcp/tests/tools/test_introspection_scenarios.py
@@ -9,8 +9,8 @@ from mcp.shared.memory import (
 )
 from pipefy_sdk import PipefyClient
 
+from pipefy_mcp.core.tool_error_envelope import tool_error_message
 from pipefy_mcp.tools.introspection_tools import IntrospectionTools
-from pipefy_mcp.tools.tool_error_envelope import tool_error_message
 from tools.conftest import build_tool_test_server
