+++ packages/mcp/tests/tools/test_pipe_config_tools.py
@@ -12,6 +12,7 @@ from mcp.shared.memory import (
 )
 from pipefy_sdk import PipefyClient
 
+from pipefy_mcp.core.tool_error_envelope import tool_error, tool_error_message
 from pipefy_mcp.tools.field_condition_tools import FieldConditionTools
 from pipefy_mcp.tools.pipe_config_tool_helpers import (
     DeletePipeErrorPayload,
@@ -22,7 +23,6 @@ from pipefy_mcp.tools.pipe_config_tool_helpers import (
     normalize_phase_cards_list,
 )
 from pipefy_mcp.tools.pipe_config_tools import PipeConfigTools
-from pipefy_mcp.tools.tool_error_envelope import tool_error, tool_error_message
 from tools.conftest import assert_invalid_arguments_envelope, build_tool_test_server
