+++ packages/mcp/tests/tools/test_portal_tools.py
@@ -14,8 +14,8 @@ from mcp.shared.memory import (
 from pipefy_sdk import PipefyClient
 from pipefy_sdk.exceptions import PortalPermissionError
 
+from pipefy_mcp.core.tool_error_envelope import tool_error_message
 from pipefy_mcp.tools.portal_tools import PortalTools
-from pipefy_mcp.tools.tool_error_envelope import tool_error_message
 from tools.conftest import assert_invalid_arguments_envelope, build_tool_test_server
 
 _PORTAL_LIST_NODE = {
