+++ packages/mcp/tests/tools/test_webhook_tools.py
@@ -11,7 +11,7 @@ from mcp.shared.memory import (
 )
 from pipefy_sdk import PipefyClient
 
-from pipefy_mcp.tools.tool_error_envelope import tool_error_message
+from pipefy_mcp.core.tool_error_envelope import tool_error_message
 from pipefy_mcp.tools.webhook_tools import WebhookTools
 from tools.conftest import assert_invalid_arguments_envelope, build_tool_test_server
